+++ src/main/java/com/example/Tuned/controller/ListenerController.java
@@ -102,8 +102,10 @@ public class ListenerController {
 
         User user = userRepository.findById(listener_id).get();
         user.removeAllFollowersAndFollowing();
+        userRepository.save(user);
 
         listenerRepository.deleteById(listener_id);
-        userRepository.deleteById(listener_id);
+        if(userRepository.findById(listener_id).isPresent())
+            userRepository.deleteById(listener_id);
     }
 }
